backend/requirements.txt
flask
flask-cors
google-generativeai
gunicorn>=22.0,<24